P4TC: perf - Initial performance application
This commit contains the initial code to evaluate control path performance.
Changes were added to support two different performance values called:

update rate:
In this mode, we are creating 4 pipelines, 1 table class in each pipeline
and 32 tables in each table class and continuously updating the table
entries such that the application can measure the number of entries we
are updating into the table per one second.

latency:
In this mode, we are evaluating the latency of rule update in the table
and dump the latency values.
Following are the different modes perf_app supports:

Synchronous:
In this mode, perf_app will send the netlink message and wait for response.

Asynchronous:
In thismode, perf_app will send the netlink message and does not wait for
response.

Build procedure:
cd perf/
make clean all

Example commands:

./perf_app --rate
keeps adding the rules in P4TC tables and evaluates the rate of insertion.

./perf_app --latency
Dumps the latency of each rule update.

Signed-off-by: Gogineni, Sai Krishna <Sai.Krishna.Gogineni@intel.com>
Co-authored-by: Khan, Mohd Arif <mohd.arif.khan@intel.com>
Co-authored-by: Santosh Karthik Prasad Vajghala <santosh.karthik.prasad.vajghala@intel.com>
Co-authored-by: Pottimurthy, Sathya Narayana <sathya.narayana.pottimurthy@intel.com>